Adding the un/authorize commands
This should work, but I realized that authorizations should be per
server (and eventually per command). Everything here should work (hasn't
been tested), but I'm going to refactor the config to initialize and
store authorizations per server.

For now I'll keep the master id for development purposes but in the
future I envision a setup in which when it joins a server everyone will
be authorized to interact with the bot until explicit authorizations are
set. If the bot is removed from a server, it should clear out the
authorizations for that server. Here's why:

1) Bot is added to a server, the authorizations for that server are
initialized to be empty so everyone is authorized.
2) Someone adds users or roles to the authorization list.
3) If those users/roles were done incorrectly (gave it to someone who no
longer gets on the server for example), the bot would effectively be
unusable.
4) Remove the bot from the server. The bot sees it's been removed and
clears out all the authorizations stored for that server.
5) Bot is re-added to the server, authorizations are reset (step 1)

Also, authorization should be done at the command level and should
simply be a call to an authorization method that says "is this user
authenticated for this command". All of the authentication logic should
be in that method (given a user and a command, see if that user is
authorized directly or is part of a role that is authorized).