LORE-33: Message handler now checks authorization for commands again
Since the message handler has the command object, it can call out to the
authorization method to check the caller's permission. This saves having
to do it from each of the command methods.

Also cleaned up the top command method as there was some cruft in there
from the first version of the bot that had hard coded the suggestion
channel name.