Added a check to see if a matched command should defer to another
If a command can be legitimately mentioned in another command that comes
after it in the command list (ex: unauthorized command can accept
authorize in its message), it needs to be marked as such so that the
message handler can check to see if it needs to defer to the command
further down the list.

This deferment should be done in the message handler so that proper
permissions can be checked before executing the command.