Added requirements.txt, updated dependencies
Haven't touched this in a long while, but I spent a little bit of time getting this to the point of being able to connect to Discord again.

There is currently a problem with reading the config.yml back in with PyYAML. A work around is just to make sure config.yml doesn't exist when starting up, but that means no config settings will persist between sessions.